Make the function that creates a device token and signs in using it idempotent.
This is necessary as this function will be batched with other calls at the time of submitting the profile page in the onboarding flow.

Also, minimize the user change notifications to app state. We only need to notify the AppState/RouterDelegate of events when the user signs off *or* signs in after being signed off.